genai/aws-gen-ai-kr/13_agentcore/tutorials/06-AgentCore-observability/03-advanced-concepts/02-custom-span-agent-langgraph/requirements.txt
uv
ipykernel>=6.29.5
jupyter>=1.1.1
boto3>=1.39.9
langchain>=0.3.26
langgraph>=0.5.3
strands-agents>=1.0.1
bedrock-agentcore==0.1.1
strands-agents-tools>=0.2.1
mcp>=1.12.0
requests-toolbelt>=1.0.0
awscli>=1.41.10
termcolor>=3.1.0
langchain-aws>=0.2.29
bedrock-agentcore-starter-toolkit
uvicorn
opentelemetry-instrumentation-langchain